changed frame_t.data.int_data to float

+++ rsDaemon/rs-daemon.cpp
@@ -50,7 +50,7 @@ struct frame_t {
     long int length;
     union data_union{
         char char_data[BUF_SIZE];
-        int int_data[BUF_SIZE];
+        float int_data[BUF_SIZE];
     }data;
 };
 union received_msg {
@@ -151,7 +151,7 @@ void parse_received_msg(const uint8_t * msg, int length)
 }
 void send_depth_data(array_packet * center_coordinates)
 {
-    int * result_array  = get_depth_from_coordinates(center_coordinates);
+    float * result_array  = get_depth_from_coordinates(center_coordinates);
 
     struct frame_t frame;
     memset(&frame, 0, sizeof(frame));
@@ -269,8 +269,8 @@ void metadata_to_csv(const rs2::frame& frm, const std::string& filename)
     csv.close();
 }
 
-int * get_depth_from_coordinates(array_packet * center_coordinates){
-    int * coordinate_list = new int[center_coordinates->len];
+float * get_depth_from_coordinates(array_packet * center_coordinates){
+    float * coordinate_list = new int[center_coordinates->len];
     memcpy(coordinate_list,center_coordinates->data, center_coordinates->len* sizeof(int));
     size_t list_len = center_coordinates->len;
